requirements.txt
pyaudio
streamlit-audiorec
matplotlib
sounddevice
torch
torchaudio
scipy
pydub
librosa